requirements.txt
watchdog
faster-whisper
anthropic
openai
jinja2
pyyaml
python-frontmatter
